+++ apps/main/src/llamalend/LlamaMarketsPage/cells/CompactUsdCell.tsx
