+++ apps/main/src/llamalend/LlamaMarketsPage/cells/RateCell/RewardsIcons.tsx
