+++ apps/main/src/llamalend/LlamaMarketsPage/cells/RateCell/RewardsTooltipItems.tsx
