+++ apps/main/src/llamalend/LlamaMarketsPage/header-tooltips/BorrowRateHeaderTooltipContent.tsx
