+++ apps/main/src/llamalend/LlamaMarketsPage/header-tooltips/CollateralBorrowHeaderTooltipContent.tsx
