+++ apps/main/src/llamalend/LlamaMarketsPage/header-tooltips/LendRateHeaderTooltipContent.tsx
